Java project illustrating a traffic plan
This project and plan identify the key points of the purpose, algorithm, and process that went into completing the requirements to finish. The overall goal of this assignment was to create a vehicle object that moves across a designated pathway, reaches the other side, and then moves back. As seen in the code, this was completed and executed perfectly, where the method used consisted of an if() block, that sets parameters if the value is inside the threshold. If the value is in fact inside of the threshold, then the stepSize = -stepSize (or the inverse) and if not, then the value of stepSize is added and declared back to the xOffset variable.